+++ src/Cwaff-Guns/AimuHakurei.cs
@@ -198,6 +198,8 @@ public class AimuHakurei : CwaffGun
     public override void OwnedUpdatePlayer(PlayerController player, GunInventory inventory)
     {
         base.OwnedUpdatePlayer(player, inventory);
+        if (!this.gun)
+            return; // can happen with Paradox in Breach or via Expand's gunball machine
         UpdateGraze();
     }