+++ src/Cwaff-Guns/Alyx.cs
@@ -58,6 +58,8 @@ public class Alyx : CwaffGun
     public override void OwnedUpdatePlayer(PlayerController player, GunInventory inventory)
     {
         base.OwnedUpdatePlayer(player, inventory);
+        if (!this.gun)
+            return; // can happen with Paradox in Breach or via Expand's gunball machine
         RecalculateAmmo();
     }