+++ changelog.md
@@ -7,6 +7,7 @@
 - Fixed null deref caused by Paintball Cannon trying to apply its affects to certain non-Paintball Cannon projectiles
 - Added some temporary patches to fix a handful of compatibility issues with JuneLib until it gets properly updated
 - Increased the size of Stereoscope's visual indicators and added extra wedge showing current aim direction for increased clarity
+- Fixed null derefs caused by certain guns trying to call update methods in the Breach when given via Expand the Gungeon's gunball machine
 
 ## 1.28.1 (2025-10-01)